Relocate default domain lookup under common NS
