+++ classes/external.php
@@ -13,7 +13,7 @@ use external_api;
 use external_function_parameters;
 use external_single_structure;
 use external_value;
-use local_signin\model\user_default_domain;
+use local_signin\domainfinder\user_default_domain;
 
 defined('MOODLE_INTERNAL') || die;
